+++ src/main/java/de/samply/annotations/RoleConstraints.java
@@ -1,7 +1,5 @@
 package de.samply.annotations;
 
-import de.samply.project.ProjectBridgeheadState;
-import de.samply.project.ProjectState;
 import de.samply.user.OrganisationRole;
 import de.samply.user.ProjectRole;
 
@@ -11,7 +9,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
+@Target(ElementType.METHOD)
 public @interface RoleConstraints {
 
     OrganisationRole[] organisationRoles = null;