+++ src/main/java/de/samply/annotations/StateConstraints.java
@@ -1,7 +1,7 @@
 package de.samply.annotations;
 
-import de.samply.project.ProjectBridgeheadState;
-import de.samply.project.ProjectState;
+import de.samply.project.state.ProjectBridgeheadState;
+import de.samply.project.state.ProjectState;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,7 +9,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
+@Target(ElementType.METHOD)
 
 public @interface StateConstraints {
     ProjectState[] projectStates = null;