+++ src/main/java/de/samply/bridgehead/BridgeheadConfiguration.java
@@ -1,0 +1,20 @@
+package de.samply.bridgehead;
+
+import de.samply.app.ProjectManagerConst;
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.List;
+
+@Configuration
+@Getter
+public class BridgeheadConfiguration {
+
+    private List<String> bridgeheads;
+
+    public BridgeheadConfiguration(@Value(ProjectManagerConst.BRIDGEHEADS_SV) List<String> bridgeheads) {
+        this.bridgeheads = bridgeheads;
+    }
+
+}