+++ CHANGELOG.md
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Session User Info email
 - Multi bridgehead user
 - Test user if security disabled
+- Bridgehead configuration